requirements.txt
veadk-python
pydantic
python-dotenv
